refactor: move sms components to a sms folder
