+++ src/components/app/pageHome/common/hero/heroCTA.tsx
@@ -2,7 +2,7 @@
 import { SMSStatus } from '@prisma/client'
 
 import { LoginDialogWrapper } from '@/components/app/authentication/loginDialogWrapper'
-import { SMSOptInCTA } from '@/components/app/smsOptInCTA'
+import { SMSOptInCTA } from '@/components/app/sms/smsOptInCTA'
 import { OPEN_UPDATE_USER_PROFILE_FORM_QUERY_PARAM_KEY } from '@/components/app/updateUserProfileForm/queryParamConfig'
 import { Button } from '@/components/ui/button'
 import { InternalLink } from '@/components/ui/link'
